requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
django-taggit==6.1.0
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
